Refactor for Codacy conformance
